修正MinGW编译失败
git-svn-id: http://qlanguage.googlecode.com/svn/trunk@97 6d982bb3-a5a7-393d-ffc5-582415be9c85

+++ QCoreUnitTest/Units/TestRegex.cpp
@@ -120,5 +120,5 @@ TEST_CASE(TestRegex)
     }
     t = clock() - t;
     allocator<char>::deallocate(szTest, iContentSize);
-    printf("parse data length: %lu bytes\nuse of time: %fs\n", iContentSize, t / 1000.0f);
+    printf("parse data length: %lu bytes\nuse of time: %fs\n", (unsigned long)iContentSize, t / 1000.0f);
 }
